requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
gunicorn>=20.0
yfpy>=2.0
python-dotenv>=0.19
requests-oauthlib
gevent
yahoo_fantasy_api>=2.0
apscheduler
pandas
google-cloud-storage
redis
rq
pytz
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25